pyproject.toml
[project]
name = "geospot-vlm"
version = "0.1.0"
description = "Train VLMs to predict geographic coordinates from images"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
authors = [
    {name = "Surya Dantuluri"},
]
keywords = ["vlm", "geolocation", "reinforcement-learning", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "tinker",
    "transformers>=4.40.0",
    "torch>=2.0",
    "Pillow>=10.0",
    "chz>=0.4.0",
    "datasets>=2.18.0,<3.0.0",
    "huggingface-hub>=0.23.0",
    "numpy>=1.24",
    "webdataset>=1.0.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
]
webdataset = [
    "webdataset>=0.2.100",
]
wandb = [
    "wandb>=0.16.0",
]

[project.scripts]
geospot-sft = "geospot.sft:cli_main"
geospot-train = "geospot.train:main"

[project.urls]
Homepage = "https://github.com/sdan/geospot-vlm"
Repository = "https://github.com/sdan/geospot-vlm"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["geospot"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
